+++ options/options_test.go
@@ -510,6 +510,9 @@ func (s *OptUtilSuite) TestArguments(c *C) {
 	a = a.Unshift("0", "1")
 
 	c.Assert(a, DeepEquals, Arguments{"0", "1", "2", "3", "4", "5"})
+
+	a = Arguments{"*.txt", "*.jpg"}
+	c.Assert(a.Filter("*.txt"), HasLen, 0)
 }
 
 func (s *OptUtilSuite) TestArgumentsConvertion(c *C) {
@@ -523,6 +526,21 @@ func (s *OptUtilSuite) TestArgumentsConvertion(c *C) {
 	c.Assert(a.Get(2).String(), Equals, "2.67")
 	c.Assert(a.Get(3).String(), Equals, "true")
 	c.Assert(a.Get(4).String(), Equals, "")
+
+	c.Assert(a.Get(0).Is("test"), Equals, true)
+	c.Assert(a.Get(0).Is("abcd"), Equals, false)
+	c.Assert(a.Get(1).Is(6), Equals, true)
+	c.Assert(a.Get(1).Is(12), Equals, false)
+	c.Assert(a.Get(1).Is(int64(6)), Equals, true)
+	c.Assert(a.Get(1).Is(int64(12)), Equals, false)
+	c.Assert(a.Get(1).Is(uint64(6)), Equals, true)
+	c.Assert(a.Get(1).Is(uint64(12)), Equals, false)
+	c.Assert(a.Get(2).Is(2.67), Equals, true)
+	c.Assert(a.Get(2).Is(3.14), Equals, false)
+	c.Assert(a.Get(3).Is(true), Equals, true)
+	c.Assert(a.Get(3).Is(false), Equals, false)
+
+	c.Assert(a.Get(0).Is([]string{}), Equals, false)
 
 	vi, err := a.Get(0).Int()
 	c.Assert(err, NotNil)