+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 12.49.0
+
+* `[options]` Added method `Is` for checking argument value
+* `[options]` Fix bug with filtering unexpanded globs
+
 ### 12.48.0
 
 * `[terminal]` Added option HideLength for hiding password length